ai-python/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-groq
langchain-community
langchain-openai
huggingface_hub
requests
python-multipart
sentence-transformers
